+++ sheets/PersonSheet.js
@@ -143,10 +143,10 @@ export class PersonSheet extends EnhancedJournalSheet {
     }
 
     async addRelationship(data) {
-        let relationship = await this.getEntity(data);
+        let relationship = await this.getDocument(data);
 
-        if (relationship.entity) {
-            let type = relationship.entity.data.flags['monks-enhanced-journal']?.type
+        if (relationship.document) {
+            let type = relationship.document.data.flags['monks-enhanced-journal']?.type
             relationship.data.type = type;
             if (['organization','person','place'].includes(type)) {
                 let actors = duplicate(this.object.data.flags["monks-enhanced-journal"].actors || []);
@@ -168,9 +168,9 @@ export class PersonSheet extends EnhancedJournalSheet {
     }
 
     async addActor(data) {
-        let actor = await this.getEntity(data);
+        let actor = await this.getDocument(data);
 
-        if (actor.entity) {
+        if (actor.document) {
             this.object.setFlag("monks-enhanced-journal", "actor", actor.data);
         }
     }