+++ sheets/ShopSheet.js
@@ -361,9 +361,9 @@ export class ShopSheet extends EnhancedJournalSheet {
     }
 
     async addItem(data) {
-        let item = await this.getEntity(data);
+        let item = await this.getDocument(data);
 
-        if (item.entity) {
+        if (item.document) {
             let items = duplicate(this.object.data.flags["monks-enhanced-journal"].items || []);
 
             let olditem = items.find(i => i.id == item.data.id);
@@ -439,9 +439,9 @@ export class ShopSheet extends EnhancedJournalSheet {
     }
 
     async addActor(data) {
-        let actor = await this.getEntity(data);
+        let actor = await this.getDocument(data);
 
-        if (actor.entity) {
+        if (actor.document) {
             this.object.setFlag("monks-enhanced-journal", "actor", actor.data);
         }
     }